+++ src/client/components/react/book.js
