+++ webpack.config.js
@@ -1,7 +1,7 @@
 const path = require('path')
 
 module.exports = {
-  entry: './src/index.js',
+  entry: './src/client/index.js',
   module: {
     rules: [
       {
